Add files via upload
Understanding Problem Statement:

After visualising the data and plots we can clearly see the relations between the features of the given dataset vs purchase cost. This can be easily verified by checking the Correlation matrix given the code file.


After this before we try to process out data we also need to check if our Purchase cost is skewed or not as if it is skewed we need to apply some functions like log(1+x) to remove its skewness.

As We can clearly see this is not a good distribution because of the random increase and decrease but it is clearly not skewed so we can use it for computation.
Method of Analysis and Algorithm used:

Possible methods: 1. Linear Regression
				   		2. random forest regressor
				   		3. XG Boost